Only show social backend login links if they're configured.
We add an AppConfig to forsta_auth.saml so that the app is fully loaded,
registering its BackendMeta, before get_backends() is called. Without this,
get_backends() imports the BackendMeta for the first time, and the registry
changes underneath its iteration when the first BackendMeta.enabled property is
retrieved.

We also fix some social backend names, which have changed in
python-social-auth.