If creating an account from a social login, explicitly set an unusable password
Previously, the password was being set to '', which results in
user.has_usable_password() returning True, which in turn allows the user to
disconnect their last social auth, making it impossible for them to log in.

Now, attempting to disconnect their last social auth raises an exception if
they don't have a password set.