Discard social partial if user aborts a signup
If the user attempts a social login for an unrecognised external account,
they're passed to the signup view to onboard. If they then enter the signup
flow again by clicking on the 'Register' link, then they should no longer be
attempting to create an account with a linked social auth.

This adds a new parameter ('?from-social') to the signup view that, if missing,
ensures that the signup view doesn't attempt to link a social account.